+++ sampledb/logic/background_tasks/core.py
@@ -24,6 +24,7 @@ from .. import errors
 from .background_dataverse_export import handle_dataverse_export_task
 from .send_mail import handle_send_mail_task
 from .poke_components import handle_poke_components_task
+from .trigger_webhooks import handle_trigger_object_log_webhooks, handle_webhook_send
 
 TASK_WAIT_TIMEOUT = 30
 NUM_HANDLER_THREADS = 4
@@ -32,6 +33,8 @@ HANDLERS: typing.Dict[str, typing.Callable[[typing.Dict[str, typing.Any], typing
     'send_mail': handle_send_mail_task,
     'dataverse_export': handle_dataverse_export_task,
     'poke_components': handle_poke_components_task,
+    'trigger_object_log_webhooks': handle_trigger_object_log_webhooks,
+    'webhook_send': handle_webhook_send,
 }
 
 should_stop = False