+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Currently in development.
 - Show object form errors as expandable alert message
 - Added file name search
 - Added object_log API endpoint
+- Added object log webhooks
 
 Version 0.24.1
 --------------